pyproject.toml
[project]
name = "mchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.3.1",
    "channels-redis>=4.3.0",
    "cryptography>=46.0.2",
    "daphne>=4.2.1",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-fernet-fields>=0.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "pillow>=11.3.0",
]
